moving utils into U19 folder
